core/requirement.txt
fastapi
uvicorn
pymongo
python-multipart
python-jose[cryptography]
passlib[bcrypt]
certifi